requirements.txt
streamlit==1.28.0
pandas==2.1.0
openpyxl==3.1.2
pydantic==2.4.0
pyyaml==6.0.1
python-multipart==0.0.6
numpy==1.24.0